Add Incremental convex hull 3D
Implemented the incremental convex hull algorithm, now in 3D. Also added some additional structs such as edges and faces. My implementation works with triangle faces; this may mean there is potentially more detail that needed, but it should be fine (also it makes it much easier for me to generate a mesh to display, since I don't need to triangulate it).