Remove use of extreme points in convex hull 3D
Using extreme points when building the convex hull was used as a way to optimise the algorithm somewhat (since an initial hull with extreme points would likely remove a lot of points that are inside the convex hull). This, however, also caused issues, for example when the number of extreme points was less than 4 (e.g. (0,0,0) and (1,1,1) only two extreme points), this would cause the algorithm to crash.